Resolve BO-25 "Fix  11 24"

+++ src/test/java/tests/vindex_backoffice_ui_tests/investigationTool/trading/TradingOpenPositionsTest.java
@@ -179,8 +179,8 @@ class TradingOpenPositionsTest extends TestBaseWeb {
         openPositions.navigateOpenPositions(client.getUcid());
         openPositions.openPositionsRenders();
 
-        int from = 4;
-        int to = 8;
+        double from = 4;
+        double to = 8;
         tradingPage.openFilter();
         tradingPage.fillVolumeLotValues(String.valueOf(from), String.valueOf(to));
         tradingPage.clickApplyButton();
